Add eslint, prettier, typescript, move all code to src folder
