+++ .eslintrc.js
@@ -1,0 +1,18 @@
+module.exports = {
+  root: true,
+  parser: "@typescript-eslint/parser",
+  plugins: ["@typescript-eslint"],
+  extends: [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/eslint-recommended",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:prettier/recommended"
+  ],
+  env: {
+    node: true,
+    commonjs: true
+  },
+  rules: {
+    "@typescript-eslint/explicit-function-return-type": "off"
+  }
+};
